wildfire_crisis_demo/pyproject.toml
[project]
name = "wildfire-crisis-demo"
version = "0.1.0"
description = "Wildfire Crisis Response Coordination Agent — multi-agent decision support demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "azure-ai-projects>=1.0.0b5",
    "azure-identity>=1.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
observability = [
    "azure-monitor-opentelemetry>=1.3",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
